% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{freq_table}
\alias{freq_table}
\title{Calculate frequency of each unique value}
\usage{
freq_table(data)
}
\arguments{
\item{data}{A vector (numeric, character, or factor)}
}
\value{
A table of frequencies for each unique value
}
\description{
This function takes a vector and returns the frequency count of each unique value.
}
\examples{
freq_table(c(1, 1, 2, 3, 3, 3, 4))
}
