% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{prob_table}
\alias{prob_table}
\title{Calculate probability of each unique value}
\usage{
prob_table(data)
}
\arguments{
\item{data}{A vector (numeric, character, or factor)}
}
\value{
A table of probabilities for each unique value
}
\description{
This function takes a vector and returns the probability (relative frequency)
of each unique value. Probability = frequency / total count
}
\examples{
prob_table(c(1, 1, 2, 3, 3, 3, 4))
}
