% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Calculate Shannon Entropy}
\usage{
shannon_entropy(data, base = exp(1))
}
\arguments{
\item{data}{A vector (numeric, character, or factor)}

\item{base}{Logarithm base (default = exp(1) for natural log)}
}
\value{
Shannon entropy value
}
\description{
This function calculates the Shannon entropy of a vector.
Shannon entropy = -sum(p * log(p)) with convention 0*log(0)=0
}
\examples{
shannon_entropy(c(1, 1, 2, 3, 3, 3, 4))
shannon_entropy(c(1, 1, 2, 3, 3, 3, 4), base = 2)
}
